refactor: extract text sanitization from gateway/run.py to gateway/text_sanitizer.py
